(feat) renamed to Main-servertest.js, setup complete
